+++ src/main/kotlin/com/github/etkachev/nxwebstorm/actions/Generate.kt
@@ -1,5 +1,7 @@
 package com.github.etkachev.nxwebstorm.actions
 
+import com.github.etkachev.nxwebstorm.models.SchematicCommandData
+import com.github.etkachev.nxwebstorm.models.mapSchematicTypeStringToEnum
 import com.github.etkachev.nxwebstorm.services.MyProjectService
 import com.github.etkachev.nxwebstorm.ui.RunSchematicDialog
 import com.github.etkachev.nxwebstorm.ui.RunTerminalWindow
@@ -24,14 +26,19 @@ class Generate : AnAction() {
     val ok = dialog.showAndGet()
     if (ok && dialog.schematicSelection.containsKey("id")) {
       val id = dialog.schematicSelection["id"] ?: ""
-      val type = dialog.schematicSelection["type"] ?: ""
+      val collection = dialog.schematicSelection["collection"] ?: ""
       val fileLocation = dialog.schematicSelection["file"] ?: ""
-      val formDialog = RunSchematicDialog(proj, type, id, fileLocation)
+      val type = dialog.schematicSelection["type"] ?: ""
+      val collectionPath = dialog.schematicSelection["collectionPath"] ?: ""
+      val enumType = mapSchematicTypeStringToEnum(type)
+      val formDialog = RunSchematicDialog(proj, collection, id, fileLocation, enumType, collectionPath)
       val formOk = formDialog.showAndGet()
       if (formOk) {
         val values = formDialog.formMap.formVal
         val projectType = nxService.nxProjectType
-        val command = getSchematicCommandFromValues(type, id, values, projectType, false)
+        val schematicCommandData = SchematicCommandData(projectType, enumType, collectionPath)
+        val command =
+          getSchematicCommandFromValues(collection, id, values, schematicCommandData, false)
         val terminal = RunTerminalWindow(proj, "Run")
         terminal.runAndShow(command)
       }