+++ src/main/kotlin/com/github/etkachev/nxwebstorm/ui/RunSchematicDialog.kt
@@ -1,7 +1,9 @@
 package com.github.etkachev.nxwebstorm.ui
 
 import com.github.etkachev.nxwebstorm.actionlisteners.DryRunAction
+import com.github.etkachev.nxwebstorm.models.DryRunButtonData
 import com.github.etkachev.nxwebstorm.models.FormValueMap
+import com.github.etkachev.nxwebstorm.models.SchematicTypeEnum
 import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.openapi.project.Project
 import javax.swing.Action
@@ -9,9 +11,11 @@ import javax.swing.JComponent
 
 class RunSchematicDialog(
   private val project: Project,
-  private val type: String,
+  private val collection: String,
   private val id: String,
-  private val schematicLocation: String
+  private val schematicLocation: String,
+  private val type: SchematicTypeEnum,
+  private val collectionPath: String
 ) :
   DialogWrapper(project) {
   var formMap: FormValueMap = FormValueMap()
@@ -21,7 +25,8 @@ class RunSchematicDialog(
   }
 
   override fun createLeftSideActions(): Array<Action> {
-    return arrayOf(DryRunAction(project, type, id, formMap, this))
+    val runData = DryRunButtonData(collection, id, formMap, type, collectionPath)
+    return arrayOf(DryRunAction(project, runData, this))
   }
 
   override fun createCenterPanel(): JComponent? {