+++ src/main/kotlin/com/github/etkachev/nxwebstorm/utils/GetSchematicList.kt
@@ -2,33 +2,45 @@ package com.github.etkachev.nxwebstorm.utils
 
 import com.github.etkachev.nxwebstorm.models.FullSchematicInfo
 import com.github.etkachev.nxwebstorm.models.SchematicInfo
+import com.github.etkachev.nxwebstorm.models.SchematicTypeData
+import com.github.etkachev.nxwebstorm.models.SchematicTypeEnum
 import com.github.etkachev.nxwebstorm.models.SplitSchematicId
+import com.github.etkachev.nxwebstorm.models.mapSchematicTypeStringToEnum
+import com.intellij.ui.table.JBTable
 
 fun splitSchematicId(id: String): SplitSchematicId? {
   val split = id.split("--SPLIT--")
-  if (split.count() != 2) {
+  if (split.count() != 3) {
     return null
   }
-  val type = split[0]
+  val collection = split[0]
   val schematicId = split[1]
-  return SplitSchematicId(type, schematicId)
+  val type = split[2]
+  return SplitSchematicId(collection, schematicId, type)
 }
 
-fun findFullSchematicIdByTypeAndId(type: String, id: String, schematics: Map<String, SchematicInfo>): String? {
+fun findFullSchematicIdByTypeAndId(
+  collection: String,
+  id: String,
+  schematics: Map<String, SchematicInfo>,
+  type: SchematicTypeEnum
+): String? {
+  val toMatch = generateUniqueSchematicKey(collection, id, type)
   return schematics.keys.find { key ->
-    val toMatch = generateUniqueSchematicKey(type, id)
     key == toMatch
   }
 }
 
-fun generateUniqueSchematicKey(type: String, id: String): String {
-  return "$type--SPLIT--$id"
+fun generateUniqueSchematicKey(collection: String, id: String, type: SchematicTypeEnum): String {
+  val typeData = type.data
+  return "$collection--SPLIT--$id--SPLIT--$typeData"
 }
 
 fun getSchematicData(schematics: Map<String, SchematicInfo>): Array<FullSchematicInfo> {
   return schematics.keys.mapNotNull { id ->
     val splitData = splitSchematicId(id) ?: return@mapNotNull null
-    val type = splitData.type
+    val collection = splitData.collection
+    val type = SchematicTypeData(collection, splitData.type)
     val schematicId = splitData.id
     val description = schematics[id]?.description
     val fileLocation = schematics[id]?.fileLocation ?: ""
@@ -44,4 +56,12 @@ fun flattenMultipleMaps(vararg maps: Map<String, SchematicInfo>): Map<String, Sc
     }
     return@fold acc
   }).toMap()
+}
+
+fun getSchematicIdFromTableSelect(table: JBTable, selectedRow: Int, schematics: Map<String, SchematicInfo>): String? {
+  val collection = table.getValueAt(selectedRow, 0).toString()
+  val id = table.getValueAt(selectedRow, 1).toString()
+  val type = table.model.getValueAt(table.convertRowIndexToModel(selectedRow), 3).toString()
+  val enumType = mapSchematicTypeStringToEnum(type)
+  return findFullSchematicIdByTypeAndId(collection, id, schematics, enumType)
 }